Conflicts:
	app/scripts/controllers/reports.js

+++ NMSReportingSuite/src/main/java/com/beehyv/nmsreporting/model/FlwImportRejection.java
@@ -16,10 +16,10 @@ public class FlwImportRejection {
     private Long id;
 
     @Column(name = "state_id")
-    private Long stateId;
+    private Integer stateId;
 
     @Column(name = "district_id")
-    private Long districtId;
+    private Integer districtId;
 
     @Column(name = "district_name")
     private String districtName;
@@ -31,7 +31,7 @@ public class FlwImportRejection {
     private String talukaName;
 
     @Column(name = "health_block_id")
-    private Long healthBlockId;
+    private Integer healthBlockId;
 
     @Column(name = "health_block_name")
     private String healthBlockName;
@@ -164,19 +164,19 @@ public class FlwImportRejection {
         this.id = id;
     }
 
-    public Long getStateId() {
+    public Integer getStateId() {
         return stateId;
     }
 
-    public void setStateId(Long stateId) {
+    public void setStateId(Integer stateId) {
         this.stateId = stateId;
     }
 
-    public Long getDistrictId() {
+    public Integer getDistrictId() {
         return districtId;
     }
 
-    public void setDistrictId(Long districtId) {
+    public void setDistrictId(Integer districtId) {
         this.districtId = districtId;
     }
 
@@ -204,11 +204,11 @@ public class FlwImportRejection {
         this.talukaName = talukaName;
     }
 
-    public Long getHealthBlockId() {
+    public Integer getHealthBlockId() {
         return healthBlockId;
     }
 
-    public void setHealthBlockId(Long healthBlockId) {
+    public void setHealthBlockId(Integer healthBlockId) {
         this.healthBlockId = healthBlockId;
     }
